Separate exception hooks from message helpers
Move all of the message rendering and fixed strings into a dedicated
`messages` module, and take a pass over the resulting hooks module to
try to make things more compact and readable.
